+++ tests/actors/shading.py
@@ -886,16 +886,8 @@ class TestResetAllManualHand(tests.helper.test_case_base.TestCaseBase):
 
 		self.reset_shading_rule = habapp_rules.actors.shading.ResetAllManualHand("Unittest_Reset")
 
-	def test_init(self):
-		"""Test __init__"""
-		with unittest.mock.patch("HABApp.rule.scheduler.habappschedulerview.HABAppSchedulerView.soon") as run_soon_mock:
-			reset_shading_rule = habapp_rules.actors.shading.ResetAllManualHand("Unittest_Reset")
-		run_soon_mock.assert_called_once_with(reset_shading_rule._ResetAllManualHand__get_shading_objects)
-
 	def test__get_shading_objects(self):
 		"""Test __get_shading_objects."""
-		self.assertEqual([], self.reset_shading_rule._ResetAllManualHand__shading_objects)
-
 		with unittest.mock.patch.object(self.reset_shading_rule, "get_rule") as get_rule_mock:
 			self.reset_shading_rule._ResetAllManualHand__get_shading_objects()
 
@@ -918,16 +910,15 @@ class TestResetAllManualHand(tests.helper.test_case_base.TestCaseBase):
 		shading_rule_mock = unittest.mock.MagicMock(spec=habapp_rules.actors.shading.Raffstore)
 		shading_rule_mock._item_manual = unittest.mock.MagicMock(spec=HABApp.openhab.items.SwitchItem)
 
-		self.reset_shading_rule._ResetAllManualHand__shading_objects = [shading_rule_mock]
+		with unittest.mock.patch.object(self.reset_shading_rule, "_ResetAllManualHand__get_shading_objects", return_value=[shading_rule_mock]):
+			for test_case in test_cases:
+				shading_rule_mock.state = test_case.state
+				shading_rule_mock._item_manual.oh_send_command.reset_mock()
 
-		for test_case in test_cases:
-			shading_rule_mock.state = test_case.state
-			shading_rule_mock._item_manual.oh_send_command.reset_mock()
+				self.reset_shading_rule._cb_reset_all(HABApp.openhab.events.ItemCommandEvent("name", test_case.event))
 
-			self.reset_shading_rule._cb_reset_all(HABApp.openhab.events.ItemCommandEvent("name", test_case.event))
-
-			self.assertEqual(len(test_case.manual_commands), shading_rule_mock._item_manual.oh_send_command.call_count)
-			shading_rule_mock._item_manual.oh_send_command.assert_has_calls(test_case.manual_commands)
+				self.assertEqual(len(test_case.manual_commands), shading_rule_mock._item_manual.oh_send_command.call_count)
+				shading_rule_mock._item_manual.oh_send_command.assert_has_calls(test_case.manual_commands)
 
 
 class TestSlatValueSun(tests.helper.test_case_base.TestCaseBase):