+++ changelog.md
@@ -4,6 +4,10 @@
 
 - added rule ``habapp_rules.system.sleep.LinkSleep`` to link multiple sleep rules
 
+## Bugfix
+
+- fixed bug in ``habapp_rules.actors.shading.ResetAllManualHand`` which did not always reset all shading instances.
+
 # Version 5.1.0 - 06.10.2023
 
 ## Features
@@ -12,7 +16,7 @@
 
 ## Bugfix
 
-- fixed bug in ``habapp_rules.actors.schading._ShadingBase`` which caused a switch to night close if it was not configured.
+- fixed bug in ``habapp_rules.actors.shading._ShadingBase`` which caused a switch to night close if it was not configured.
 
 # Version 5.0.0 - 01.10.2023